requirements.txt
discord.py
praw
youtube-dl==2021.6.6
